requirements/base.txt
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==2.2.8
django-filter==2.3.0
django-ses==1.0.3
djangorestframework==3.12.2
djangorestframework_simplejwt==4.6.0
graphene-django==2.15.0
markdown>=3.0,<=3.3.4
mysqlclient==2.0.1
boto3>=1.15.0,<=1.15.3
graphene-file-upload==1.3.0
django-graphql-jwt==0.3.4